refactor(prompts): decompose prompt-definitions into prompts/ directory
Mirrors the established tools/ decomposition pattern. The 736-line
monolith is now a 47-line orchestrator that delegates to three group
modules under prompts/, each with its own PROMPT_NAMES constant and
register function. Shared helpers (PromptRegistrationContext type,
textResult, wrapWithDataMarkers, maxCharsArg) live in prompt-helpers.ts.

The 1439-line test file is split into a shared harness + three per-prompt
test files + a slimmed orchestrator test. Same 66 tests, same public API
(mcp-router.ts unchanged), no behavioral changes.
